* when control function created, it is bound to this, the current instance of audio-control
This means that you can reference properties from this.parentElement (the one your controlling), or anywhere else you can get to from this.

Its a bit slow now due to the fact that two-way data binding is used such that poking values into the element itself changes the UI, and changing the UI changes properties of the element.