** xtc.html:
This seems to work very well!! It's the closest I've ever come to getting crosstalk cancelation to sound convincing. Little coloration, and it really does excentuate perception of spacial cues, especially on true binaural sources.

class XTC:

trying to encode xtc.html into a low level component... Results are not the same as for xtc.html, and I do not know why!!