* test now outlines a much simpler way of insuring elements can connect themselves and run their property effects at the proper time (i.e. only after they are ready / connected)
in elements whose children need to be ready before they can function, call waitForChildren() in connectedCallback with arguments element, and a callback which runs when that element becomes ready. The callback can then operate on the children and when finished, signal that it itself is ready. These events then propagate up the tree until audio-context becomes ready.

we can also call runPropertyEffects() with element as arg to run all property observers set up in the static properties object. This should be called from within the waitForChildren() callback.